Implement -reject/allow commands
Rename command stub fucntion / command name deny -> reject. This is
consistent with the way that the rest of the project describes not
verifying some candidate. An alias has been added to preserve the
command calling

Merge the handlers into one function handle, which is called by each
command with an enum option. This is good DRY!

Remove cleanup/gate, since handle was merged